Run multiple lighthouse
